fix(controller): split repo by '/'
This covers all cases of acceptable docker image names. Since namespaces
with '.' in the name is an invalid character, if we tokenize the name
and we still find a '.' in the namespace, it means it has a hostname.

Checking for '/' is necessary for images which are not namespaced, such
as 'deis.is.awesome'.